ulong instead of uint; patch by chrissalch CCBUG:207327
svn path=/branches/KDE/4.3/kdeplasma-addons/; revision=1029410

+++ applets/microblog/microblog.h
@@ -162,10 +162,10 @@ class MicroBlog : public Plasma::PopupApplet
         QString m_curTimeline;
 
         QMap<QString, QPixmap> m_pictureMap;
-        QMap<uint, Plasma::DataEngine::Data> m_tweetMap;
+        QMap<ulong, Plasma::DataEngine::Data> m_tweetMap;
         QList<Tweet> m_tweetWidgets;
 
-        uint m_lastTweet;
+        ulong m_lastTweet;
         KWallet::Wallet *m_wallet;
         enum WalletWait { None=0, Read, Write };
         WalletWait m_walletWait;
